[Refactor] Move and restructure Ranking components
