+++ src/components/layouts/Ranking/Children/Global/GlobalTable/PlayerRow.tsx
