+++ src/components/layouts/Ranking/Children/Global/GlobalTable/SearchPlayer.tsx
