+++ src/components/layouts/Ranking/Children/Global/GlobalTable/usersData.ts
