+++ src/pages/Ranking/Children/Clan.tsx
